Add files via upload
I added some new backgrounds to match the drip of the game, and readjusted everything to make sure it is neat. I also added music that is actually for the ears, but I will add an option later on for the user to choose the music that they would want so it can be fun.