Add files via upload
I made my music option screen fully done (I forgot to save it😭), but it now has two music options. I also fully now made my geo sprite to move when the user presses on the space bar by its x-position. I am now trying to make it better and try to make the obstacles appear randomly as when it goes off the screen with a little extra small things as the main problem is done.