Add files via upload
I added a timer thing on the screen just to have it as a feature. I was trying to figure out the error in my code of where if you click the space while the user looses it does not work and try implementing it with my other geo sprites.